Add BAKE_AND_REHEAT_CALLS to replay.js.
* Also fix player.html if LogCanvas web-ext isn't actually loaded. :)
* Report (re)play time in console.

Baking and reheating seems to run into trouble with a 20k-fish aquarium
recording, whereas with it disabled, it manages ok. (To be fair, it's a
13M call, 800MB json recording)

While I kept it disable-able, what's really happening is it's building
and running the baked version, just not saving it to be run again.

What happens is it gets ~52-56 frames in to replay before having a frame
of incomplete rendering and hanging/stalling silently.

In about:memory:

> 124.27 MB (02.64%) - string(length=9, copies=5429600, "uniform1f")/gc-heap/latin1
> 123.57 MB (02.63%) - string(length=10, copies=5398791, "uniform3fv")/gc-heap/latin1
>  62.12 MB (01.32%) - string(length=12, copies=2714163, "drawElements")/gc-heap/latin1
>  61.15 MB (01.30%) - string(length=4, copies=2671851, "$431")/gc-heap/latin1
>  61.15 MB (01.30%) - string(length=4, copies=2671851, "$433")/gc-heap/latin1
>  61.15 MB (01.30%) - string(length=4, copies=2671851, "$435")/gc-heap/latin1
>  61.15 MB (01.30%) - string(length=4, copies=2671851, "$437")/gc-heap/latin1

You can see the rough breakdown of calls by type from the millions of
strings!
While a couple hundred MB of strings is survivable, probably making
a couple non-trivial baking/reheat closures for each of the 13M calls
isn't viable.
Fortunately, I bet I can dedupe a heck of a lot of these!